% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_utils.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directories if not exist}
\usage{
create_dirs(dir, remove.dir = FALSE)
}
\arguments{
\item{dir}{a character vector of path names;}

\item{remove.dir}{a logical value. If FALSE, only the path not exist be created.}
}
\value{
a path
}
\description{
Create directories if not exist
}
\examples{
create_dirs("tmp_dir")
}
\concept{utils}
