% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_utils.R
\name{dims}
\alias{dims}
\title{<dim> for many matrices}
\usage{
dims(mats)
}
\arguments{
\item{mats}{a list of matrices (or a single matrix)}
}
\value{
dim for each matrix provided.
}
\description{
Returns the result of dim for every matrix in a list
}
\concept{utils}
