% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_utils.R
\name{time_str}
\alias{time_str}
\title{A format time output as a string character}
\usage{
time_str(ymd = T, hms = T)
}
\arguments{
\item{ymd}{a logical value}

\item{hms}{a logical value}
}
\value{
a format time
}
\description{
if only ymd set TRUE, year-month-day
if only hms set TRUE, hour minute second
if all TRUE, year_month_day_hour minute second
}
\examples{
time_str()
}
\concept{utils}
